Move ZIP-related things to own dir
